+++ packages/fs/tsdown.config.ts
@@ -1,0 +1,17 @@
+import { defineConfig, type UserConfig } from 'tsdown';
+
+const config: UserConfig = defineConfig([
+  {
+    entry: ['src/**/*.ts'],
+    tsconfig: './tsconfig.json',
+    clean: true,
+    dts: {
+      sourcemap: true,
+      tsgo: true,
+    },
+    unbundle: true,
+    platform: 'neutral',
+  },
+]);
+
+export default config;